progress to date. added comments and trying to solve the mutex problem. still in progress.

+++ commands/VendorLiquidationController.php
@@ -16,26 +16,38 @@ use app\modules\checkout\models\search\PaymentSearch;
 class VendorLiquidationController extends Controller{
 
     public function actionLiquidationByLot(){
-        // todo: implement MUTEX library for this controller. So as to run only if no other liquidation is running ...
-
         $this->stdout("\nLiquidation By Lot initiated.\n"); // you can catch this output by using >> as an extra parameter via command line interface
         
         $model = VendorLiquidationProcess::find()->where(['status' => 'pending'])->one();
-        $this->LiquidationProcess($model);
 
-        $model->status = 'success';
-        // todo: dont force the saving. so as not to break anything in the future that might rely on 'perfect' data structures
-        $model->save(false);
-        /* if($model->save()){
-            do ...
-        } else {
-            error catch and summary() ...
-        } */
-        
-        $vendor_liquidations = VendorLiquidation::find()->where(['vendor_liquidation_process_id' => $model->vendor_liquidation_process_id])->all();
+        if($model) { // if liquidation ANY 'pending' vendor liquidation is found ...
+            try {
+                if(Yii::$app->mutex->acquire('mutex_liquidate_vendors')) { // only one liquidation can be active at a time
+                    $this->stdout("Vendor Liquidation Process ID: ".$model->vendor_liquidation_process_id."\n");
 
-        foreach ($vendor_liquidations as $key => $value) {
-	        if($value->status == "cancelled" || empty($value->total)) $value->delete();
+                    $this->LiquidationProcess($model);
+                    $model->status = 'success';
+                    
+                    // todo: dont force the saving. so as not to break anything in the future that might rely on 'perfect' data structures
+                    $model->save(false);
+
+                    /* if($model->save()){
+                        do ...
+                    } else {
+                        error catch and summary() ...
+                    } */
+                    
+                    $vendor_liquidations = VendorLiquidation::find()->where(['vendor_liquidation_process_id' => $model->vendor_liquidation_process_id])->all();
+            
+                    foreach ($vendor_liquidations as $key => $value) {
+                        if($value->status == "cancelled" || empty($value->total)) $value->delete();
+                    }
+                }
+            } catch (\Exception $ex) {
+                $err_msg="ERROR__________". 'Linea '.$ex->getLine()."\n" .'Archivo '.$ex->getFile() ."\n" .$ex->getMessage() ."\n" .$ex->getTraceAsString()."\n";
+                $this->stdout($err_msg);
+                //\Yii::info('ERROR ________________ ' . $ex->getMessage() ."\n" .$ex->getTraceAsString(), 'facturacion-creacion');
+            }
         }
     }
